requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20231117
google-generativeai==0.3.2
gtts==2.4.0
werkzeug==2.3.7

# Node.js dependencies (install separately):
# npm install sanskrit-tts